+++ CHANGES.md
@@ -3,8 +3,9 @@
 # dev
 
 * Print output of wheel build result.
-* `config-env`: don't require write access to FME_HOME folder.
 * Fix handling of custom transformers that have multiple categories.
+* `config-env`: don't require write access to the FME installation folder.
+* `config-env`: verify successful configuration by importing fmeobjects.
 
 # 1.16.1